requirements.txt
Django
djangorestframework
drfpasswordless-gstr169
PyJWT
